+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [N.N.N] - XXXX-XX-XX
 
+- Detect valid Codex models the right way
 - Remove stray blank line from incoming mid-round notices
 - Spell grant flags in capability order (`rxw`)
 - Warn when [sandbox.read] is used redundantly